Extract Bootcamp response mapper into separate class
